Readme.md [Online Shopping Fraud]
One of the main places where different types of scammers look for potential victims is classified advertisements. On such a forum, con artists can be found as both buyers and sellers. To conceal their genuine identity, they can go to a variety of lengths. They frequently pose as members of the Indian Army in order to dupe unsuspecting victims. Both of these fraudulent subtypes need to be on our radar.